Optimize io.ByteBuffer in NewReader
If the supplied io.Reader is already an io.ByteBuffer, no need
to wrap it with a bufferedReader.

Change the bufferedReader scratch byte buffer to be an array
instead of a slice. This allows us to save an extra allocation.

While at it, rename bufferedReader to simpleByteReader. This reader
isn't really buffered, it just exposes an extra ReadByte function.